🔒️ v2.6.12 Enforce permissions security settings for jobs and remote actions.

+++ meerschaum/api/routes/_jobs.py
@@ -32,6 +32,8 @@ from meerschaum.api import (
     no_auth,
 )
 from meerschaum.config.static import STATIC_CONFIG
+from meerschaum.core.User import is_user_allowed_to_execute
+
 
 JOBS_STDIN_MESSAGE: str = STATIC_CONFIG['api']['jobs']['stdin_message']
 JOBS_STOP_MESSAGE: str = STATIC_CONFIG['api']['jobs']['stop_message']
@@ -141,6 +143,10 @@ def create_job(
     """
     Create and start a new job.
     """
+    allowed_success, allowed_msg = is_user_allowed_to_execute(curr_user)
+    if not allowed_success:
+        return allowed_success, allowed_msg
+
     sysargs = metadata if isinstance(metadata, list) else metadata['sysargs']
     properties = metadata['properties'] if isinstance(metadata, dict) else None
     job = Job(
@@ -172,6 +178,9 @@ def delete_job(
     """
     Delete a job.
     """
+    allowed_success, allowed_msg = is_user_allowed_to_execute(curr_user)
+    if not allowed_success:
+        return allowed_success, allowed_msg
     job = _get_job(name)
     return job.delete()
 
@@ -221,6 +230,10 @@ def start_job(
     """
     Start a job if stopped.
     """
+    allowed_success, allowed_msg = is_user_allowed_to_execute(curr_user)
+    if not allowed_success:
+        return allowed_success, allowed_msg
+
     job = _get_job(name)
     if not job.exists():
         raise fastapi.HTTPException(
@@ -240,6 +253,10 @@ def stop_job(
     """
     Stop a job if running.
     """
+    allowed_success, allowed_msg = is_user_allowed_to_execute(curr_user)
+    if not allowed_success:
+        return allowed_success, allowed_msg
+
     job = _get_job(name)
     if not job.exists():
         raise fastapi.HTTPException(
@@ -259,6 +276,10 @@ def pause_job(
     """
     Pause a job if running.
     """
+    allowed_success, allowed_msg = is_user_allowed_to_execute(curr_user)
+    if not allowed_success:
+        return allowed_success, allowed_msg
+
     job = _get_job(name)
     if not job.exists():
         raise fastapi.HTTPException(
